Move and refactor app_backend clients
