+++ app_backend/tests/test_eval_task_routes.py
@@ -5,7 +5,7 @@ from fastapi import FastAPI
 from fastapi.testclient import TestClient
 from httpx import ASGITransport, AsyncClient
 
-from metta.app_backend.eval_task_client import EvalTaskClient
+from metta.app_backend.clients.eval_task_client import EvalTaskClient
 from metta.app_backend.metta_repo import MettaRepo
 from metta.app_backend.routes.eval_task_routes import (
     TaskClaimRequest,
